# Improve Mjpeg data transfer performance