--- a/rosys/vision/image_processing.py
+++ b/rosys/vision/image_processing.py
@@ -59,19 +59,17 @@ def process_ndarray_image(image: np.ndarray, rotation: ImageRotation, crop: Opti
 
 
 def remove_exif(image_data: bytes | bytearray) -> bytes:
-    exif_marker = b'\xFF\xE1'
+    EXIF_MARKER = b'\xFF\xE1'
 
     pos = 2  # Skip SOI marker
     image_data = bytearray(image_data)
     while pos < len(image_data):
-        match_start = image_data.find(exif_marker, pos)
-        if match_start != -1:
-            match_end = match_start + 2
-            length = int.from_bytes(image_data[match_end:match_end+2], byteorder='big')
-            del image_data[match_start:match_start+length+2]  # Remove EXIF segment
-            pos = match_end
-            continue
-        else:
+        match_start = image_data.find(EXIF_MARKER, pos)
+        if match_start == -1:
             break
+        match_end = match_start + 2
+        length = int.from_bytes(image_data[match_end:match_end+2], byteorder='big')
+        del image_data[match_start:match_end+length]  # Remove EXIF segment
+        pos = match_end
 
     return bytes(image_data)