This reduces the time taken to find and create an image in the incoming data stream from around 0.23 ms an average to 0.11 ms on average on my machine.
Most of that is from avoiding the call to remove_exif every time, which will of course still take ~0.1 ms each time `capture` is called.

I think using a fixed size `bytesarray` should at least remove some work from the garbage collector though.

The changes to `remove_exif` now also half its runtime on my machine (to around 50us)