@NiklasNeugebauer 2024-05-23
@falkoschindler I incorporated your changes and improved the remove_exif function